FIX: create an upload when FastImage throws an exception
FastImage might throw an exception when it isn't able to recognize a
file as being an image (ie. happens when users changes the extension
manually)

Also improved upload specs a lot

+++ app/models/upload.rb
@@ -17,7 +17,7 @@ class Upload < ActiveRecord::Base
   validates_presence_of :original_filename
 
   def thumbnail
-    @thumbnail ||= optimized_images.where(width: width, height: height).first
+    optimized_images.where(width: width, height: height).first
   end
 
   def thumbnail_url
@@ -48,23 +48,25 @@ class Upload < ActiveRecord::Base
     sha1 = Digest::SHA1.file(file.tempfile).hexdigest
     # check if the file has already been uploaded
     unless upload = Upload.where(sha1: sha1).first
+      # deal with width & heights for images
+      if SiteSetting.authorized_image?(file)
+        # retrieve image info
+        image_info = FastImage.new(file.tempfile, raise_on_failure: true)
+        # compute image aspect ratio
+        width, height = ImageSizer.resize(*image_info.size)
+        # make sure we're at the beginning of the file (FastImage is moving the pointer)
+        file.rewind
+      end
       # create a db record (so we can use the id)
       upload = Upload.create!({
         user_id: user_id,
         original_filename: file.original_filename,
         filesize: File.size(file.tempfile),
         sha1: sha1,
-        url: ""
+        url: "",
+        width: width,
+        height: height,
       })
-      # deal with width & heights for images
-      if SiteSetting.authorized_image?(file)
-        # retrieve image info
-        image_info = FastImage.new(file.tempfile, raise_on_failure: true)
-        # compute image aspect ratio
-        upload.width, upload.height = ImageSizer.resize(*image_info.size)
-        # make sure we're at the beginning of the file (FastImage is moving the pointer)
-        file.rewind
-      end
       # store the file and update its url
       upload.url = Upload.store_file(file, sha1, upload.id)
       # save the url
@@ -80,8 +82,8 @@ class Upload < ActiveRecord::Base
   end
 
   def self.remove_file(url)
-    S3.remove_file(url) if SiteSetting.enable_s3_uploads?
-    LocalStore.remove_file(url)
+    return S3.remove_file(url) if SiteSetting.enable_s3_uploads?
+    return LocalStore.remove_file(url)
   end
 
   def self.has_been_uploaded?(url)
